--- src/main/java/com/app/model/ValidatorModel.java
@@ -1,4 +1,0 @@
-package com.app.model;
-
-public class ValidatorModel {
-}
