--- src/main/java/com/app/model/menuactions/CreateItemMenuAction.java
@@ -1,16 +1,0 @@
-package com.app.model.menuactions;
-
-import com.app.controller.ItemController;
-
-public class CreateItemMenuAction extends MenuAction {
-
-  public CreateItemMenuAction(String description) {
-    super(description);
-  }
-
-  @Override
-  public void run() {
-    new ItemController().createItem();
-  }
-}
-
