--- src/main/java/com/app/model/menuactions/DeleteItemMenuAction.java
@@ -1,15 +1,0 @@
-package com.app.model.menuactions;
-
-import com.app.controller.ItemController;
-
-public class DeleteItemMenuAction extends MenuAction {
-
-  public DeleteItemMenuAction(String description) {
-    super(description);
-  }
-
-  @Override
-  public void run() {
-    new ItemController().deleteItem();
-  }
-}
