--- src/main/java/com/app/model/menuactions/DeleteOrderMenuAction.java
@@ -1,13 +1,0 @@
-package com.app.model.menuactions;
-
-public class DeleteOrderMenuAction extends MenuAction {
-
-  public DeleteOrderMenuAction(String description) {
-    super(description);
-  }
-
-  @Override
-  public void run() {
-    System.out.println("TODO: Delete order");
-  }
-}
