--- src/main/java/com/app/model/menuactions/ExitMenuAction.java
@@ -1,13 +1,0 @@
-package com.app.model.menuactions;
-
-public class ExitMenuAction extends MenuAction {
-
-  public ExitMenuAction(String description) {
-    super(false, description);
-  }
-
-  @Override
-  public void run() {
-    System.out.println("Exit"); // TODO: Move the print to view.
-  }
-}
