--- src/main/java/com/app/model/menuactions/ViewActiveOrderMenuAction.java
@@ -1,16 +1,0 @@
-package com.app.model.menuactions;
-
-import com.app.controller.OrderController;
-
-public class ViewActiveOrderMenuAction extends MenuAction {
-
-  public ViewActiveOrderMenuAction(String description) {
-    super(description);
-  }
-
-  @Override
-  public void run() {
-    new OrderController().viewOrders();
-    System.out.println("TODO: View active orders");
-  }
-}
