--- src/main/java/com/app/model/menuactions/ViewStatisticsMenuAction.java
@@ -1,13 +1,0 @@
-package com.app.model.menuactions;
-
-public class ViewStatisticsMenuAction extends MenuAction {
-
-  public ViewStatisticsMenuAction(String description) {
-    super(description);
-  }
-
-  @Override
-  public void run() {
-    System.out.println("TODO: View statistics");
-  }
-}
