+++ src/main/java/com/app/models/ValidatorModel.java
@@ -1,0 +1,4 @@
+package com.app.models;
+
+public class ValidatorModel {
+}
