+++ src/main/java/com/app/models/menuactions/DeleteOrderMenuAction.java
@@ -1,0 +1,13 @@
+package com.app.models.menuactions;
+
+public class DeleteOrderMenuAction extends MenuAction {
+
+  public DeleteOrderMenuAction(String description) {
+    super(description);
+  }
+
+  @Override
+  public void run() {
+    System.out.println("TODO: Delete order");
+  }
+}
