+++ src/main/java/com/app/models/menuactions/ExitMenuAction.java
@@ -1,0 +1,13 @@
+package com.app.models.menuactions;
+
+public class ExitMenuAction extends MenuAction {
+
+  public ExitMenuAction(String description) {
+    super(false, description);
+  }
+
+  @Override
+  public void run() {
+    System.out.println("Exit"); // TODO: Move the print to view.
+  }
+}
