+++ src/main/java/com/app/models/menuactions/ItemSubmenuMenuAction.java
@@ -1,0 +1,20 @@
+package com.app.models.menuactions;
+
+import com.app.controllers.MenuController;
+
+public class ItemSubmenuMenuAction extends MenuAction {
+
+  public ItemSubmenuMenuAction(String description) {
+    super(description);
+  }
+
+  public void run() {
+    MenuAction[] menu = {
+        new CreateItemMenuAction("Create item"),
+        new DeleteItemMenuAction("Delete item"),
+        new ExitMenuAction("Exit"),
+    };
+
+    new MenuController("Menu management", "Please input number: ", menu).run();
+  }
+}
