+++ src/main/java/com/app/models/menuactions/ViewStatisticsMenuAction.java
@@ -1,0 +1,13 @@
+package com.app.models.menuactions;
+
+public class ViewStatisticsMenuAction extends MenuAction {
+
+  public ViewStatisticsMenuAction(String description) {
+    super(description);
+  }
+
+  @Override
+  public void run() {
+    System.out.println("TODO: View statistics");
+  }
+}
