--- src/main/java/com/app/view/ItemView.java
@@ -1,9 +1,0 @@
-package com.app.view;
-
-public class ItemView {
-
-  public void printTxt(String message){
-    System.out.println(message);
-  }
-
-}
