+++ src/main/java/com/app/views/ItemView.java
@@ -1,0 +1,9 @@
+package com.app.views;
+
+public class ItemView {
+
+  public void printTxt(String message){
+    System.out.println(message);
+  }
+
+}
