--- src/test/java/com/app/model/ConfigServiceTest.java
@@ -1,33 +1,0 @@
-package com.app.model;
-
-import com.app.model.services.ConfigService;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.io.FileNotFoundException;
-
-public class ConfigServiceTest {
-  @Test
-  public void testGetInvalidPath() {
-    ConfigService configService = new ConfigService("Test");
-    String result;
-    try {
-      result = configService.getPath();
-    } catch (FileNotFoundException e) {
-      result = null;
-    }
-    Assertions.assertEquals("", result);
-  }
-
-  @Test
-  public void testGetValidPath() {
-    ConfigService configService = new ConfigService("itemDb");
-    String result;
-    try {
-      result = configService.getPath();
-    } catch (FileNotFoundException e) {
-      result = null;
-    }
-    Assertions.assertEquals("data/itemdb/items.txt", result);
-  }
-}
