--- src/test/java/com/app/model/SimpleTest.java
@@ -1,12 +1,0 @@
-package com.app.model;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class SimpleTest {
-  @Test
-  public void testAdd() {
-    Assertions.assertEquals(1, 1);
-  }
-}
-
