+++ src/test/java/com/app/models/SimpleTest.java
@@ -1,0 +1,12 @@
+package com.app.models;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+public class SimpleTest {
+  @Test
+  public void testAdd() {
+    Assertions.assertEquals(1, 1);
+  }
+}
+
